refactor: User Settings로 이동
